+++ src/main/java/taa/command/ResetCommand.java
@@ -37,7 +37,7 @@ public class ResetCommand extends Command {
 
         String userInput = ui.getUserInput(PROMPT_WITHIN_COMMAND, true);
         String message;
-        if (userInput.equals(CONFIRM_VALUE)) {
+        if (userInput != null && userInput.equals(CONFIRM_VALUE)) {
             classList.deleteAllClasses();
             storage.save(classList);
